Use record as a pointer
Split functions in files to more clear code.
